+++ app/Providers/DuskServiceProvider.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Providers;
+
+class DuskServiceProvider extends \Laravel\Dusk\DuskServiceProvider
+{
+}