+++ resources/js/components/feature/tracking/ShipmentDetailsAndTracking.vue
@@ -25,7 +25,6 @@ import TrackingMap from '@/components/feature/tracking/TrackingMap.vue'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { useCompanyFeatures } from '@/composables/helpers'
-import { useTrackShipmentQuery } from '@/composables/queries/trackShipment'
 
 const props = defineProps({
   trackingData: {
@@ -46,13 +45,18 @@ const props = defineProps({
     type: Array,
     required: true,
   },
-  useTrackShipmentQueryRefetch: {
+  onRefresh: {
     type: Function,
     required: false,
     default: () => null,
   },
+  isRefreshing: {
+    type: Boolean,
+    required: false,
+    default: false,
+  },
   lastUpdated: {
-    type: String,
+    type: [Number, String],
     required: false,
     default: null,
   },
@@ -60,12 +64,6 @@ const props = defineProps({
 
 const { companyHasFeature } = useCompanyFeatures({
   company: props.company,
-})
-
-const { refetch, dataUpdatedAt, isRefetching } = useTrackShipmentQuery({
-  trackingNumber: props.trackingData.bolNum,
-  searchOption: `bol`,
-  companyId: props.company?.pipeline_company_id ?? ``,
 })
 
 const bolNumber = computed(() => {
@@ -97,27 +95,27 @@ const numberOfPieces = computed(() => {
 
 <template>
   <div
-    v-if="dataUpdatedAt"
+    v-if="lastUpdated"
     class="mb-4 flex items-center justify-between gap-x-2 text-xs text-muted-foreground"
   >
     <p>
       <strong>Last Updated:</strong>
 
-      {{ dayjs(dataUpdatedAt).format('MMMM D, YYYY h:mm A') }}
+      {{ dayjs(lastUpdated).format('MMMM D, YYYY h:mm A') }}
     </p>
 
     <div>
       <Button
         variant="ghost"
-        :disabled="isRefetching"
+        :disabled="isRefreshing"
         size="xs"
-        @click="refetch"
+        @click="onRefresh"
       >
         <FontAwesomeIcon
           class="mr-2"
           :icon="faSync"
           fixed-width
-          :spin="isRefetching"
+          :spin="isRefreshing"
         />
 
         <span>Refresh</span>