+++ resources/js/components/feature/tracking/ShipmentDocuments.vue
@@ -68,7 +68,7 @@ const formatFileSize = (size) => {
             </TableCell>
 
             <TableCell class="uppercase">
-              {{ document.type.split('/')[1] }}
+              {{ document.type?.split('/')?.[1] ?? 'N/A' }}
             </TableCell>
 
             <TableCell>